+++ include/Pomdog/Graphics/ImmutableIndexBuffer.hpp
@@ -36,7 +36,7 @@ public:
 	ImmutableIndexBuffer(ImmutableIndexBuffer &&) = default;
 
 	ImmutableIndexBuffer(std::shared_ptr<GraphicsDevice> const& graphicsDevice,
-		IndexElementSize indexElementSize, void const* indices, std::size_t indexCount);
+		IndexElementSize indexElementSize, void const* indices, std::uint32_t indexCount);
 
 	~ImmutableIndexBuffer() override;
 
@@ -45,7 +45,7 @@ public:
 
 	///@~Japanese
 	/// @brief インデックスの個数を取得します。
-	std::size_t GetIndexCount() const override;
+	std::uint32_t GetIndexCount() const override;
 
 	///@~Japanese
 	/// @brief インデックスの要素 1 つ分のサイズを取得します。
@@ -60,7 +60,7 @@ public:
 	
 private:
 	std::unique_ptr<Details::RenderSystem::NativeIndexBuffer> nativeIndexBuffer;
-	std::size_t indexCount;
+	std::uint32_t indexCount;
 	IndexElementSize elementSize;
 };
