+++ src/RenderSystem/NativeGraphicsContext.hpp
@@ -50,13 +50,13 @@ public:
 	/// @param indexBuffer インデックスバッファを指定します。
 	/// @param indexCount インデックスの要素数を指定します。
 	virtual void DrawIndexed(PrimitiveTopology primitiveTopology,
-		std::shared_ptr<IndexBuffer> const& indexBuffer, std::size_t indexCount) = 0;
+		std::shared_ptr<IndexBuffer> const& indexBuffer, std::uint32_t indexCount) = 0;
 
 	///@~Japanese
 	/// @brief インデックスを持たない、インスタンス化されたプリミティブを描画します。
 	/// @param primitiveTopology プリミティブの位相を指定します。
 	/// @param instanceCount レンダリングするプリミティブのインスタンスの数を指定します。
-	virtual void DrawInstanced(PrimitiveTopology primitiveTopology, std::size_t instanceCount) = 0;
+	virtual void DrawInstanced(PrimitiveTopology primitiveTopology, std::uint32_t instanceCount) = 0;
 
 	///@~Japanese
 	/// @brief インデックス付けされた頂点データに基づき、インスタンス化されたプリミティブを描画します。
@@ -65,7 +65,7 @@ public:
 	/// @param indexCount インデックスの要素数を指定します。
 	/// @param instanceCount レンダリングするプリミティブのインスタンスの数を指定します。
 	virtual void DrawIndexedInstanced(PrimitiveTopology primitiveTopology,
-		std::shared_ptr<IndexBuffer> const& indexBuffer, std::size_t indexCount, std::size_t instanceCount) = 0;
+		std::shared_ptr<IndexBuffer> const& indexBuffer, std::uint32_t indexCount, std::uint32_t instanceCount) = 0;
 		
 	///@brief Retrieves the capabilities of a GraphicsContext
 	virtual GraphicsCapabilities GetCapabilities() const = 0;
